Change message type ID from int to short

+++ src/PolyMessage.Tests.Micro/ILEmitterTests/CastToTaskOfResponseTests.cs
@@ -69,7 +69,7 @@ namespace PolyMessage.Tests.Micro.ILEmitterTests
         public void ShouldThrowWhenResponseIsUnknown()
         {
             // arrange
-            int unknownResponseTypeID = -1;
+            short unknownResponseTypeID = -1;
 
             // act
             Action act = () => _castToTaskOfResponse(unknownResponseTypeID, Task.FromResult((object) new ResponseAlpha())).Wait();
